backend/apps/dashboard/urls.py

# Copyright (C) 2018 Intel Corporation
#
# SPDX-License-Identifier: MIT

from django.urls import path
from . import views

urlpatterns = [
    path('', views.DashboardView),
    path('meta', views.DashboardMeta),
]

