backend/requirements/development.txt
-r base.txt
astroid==1.6.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.3
pylint-django==0.9.4
pylint-plugin-utils==0.2.6
rope==0.11
six==1.11.0
wrapt==1.10.11
django-extensions==2.0.6
Werkzeug==0.14.1
snakeviz==0.4.2
